+++ src/cli/commands/list.rs
@@ -14,8 +14,12 @@ use crate::{
 #[derive(Args, Debug)]
 pub struct ListArgs {
     /// List updatables packages
-    #[arg(long, default_value = "false")]
+    #[arg(long, default_value = "false", conflicts_with = "active")]
     pub updatables: bool,
+
+    /// List active packages
+    #[arg(long, default_value = "false", conflicts_with = "updatables")]
+    pub active: bool,
 }
 
 impl HandleCommand for ListArgs {
@@ -27,13 +31,26 @@ impl HandleCommand for ListArgs {
         let mut packages: Vec<&InstalledPackageVersion> = register.iterate_all().collect();
         packages.sort_by_key(|a| a.package_id.to_string());
 
-        if !self.updatables {
-            display::print_grid(packages.iter().map(|p| &p.package_id).collect());
+        if self.updatables {
+            self.updatables_list(packages, &config);
 
             return;
         }
 
-        let manager = RepositoryManager::new(&config);
+        if self.active {
+            display::print_grid(register.iterate_active_packages().collect());
+
+            return;
+        }
+
+        display::print_grid(packages.iter().map(|p| &p.package_id).collect());
+    }
+}
+
+impl ListArgs {
+    /// Lists all updatable packages.
+    fn updatables_list(&self, packages: Vec<&InstalledPackageVersion>, config: &Config) {
+        let manager = RepositoryManager::new(config);
         let mut updatables = Vec::new();
         for package in packages {
             let (_, package_meta) = manager.read_package(&package.package_id.name).unwrap_or_exit(1);