+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 - Support for loading non-text external test files.
+- The `--active` flag for the list command, to list all active packages.
 
 ### Changes
 - The build system now includes a new `package-build` xtask to create a full Packit prebuild for the release.